moved function from __init__.py to entrypoint
